Move scheduler queue code into separate root level package
